Squashed commit of the following:
    use md.childCount in metadata. fix metadtatacontols declaration
    break up metadata into vue files
    change size for series row if long
    redesigned metadata page header
